Allow double size for embeds.
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@62634 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/kernel/geos/XMLBuilder.java
@@ -304,7 +304,7 @@ public class XMLBuilder {
 		sb.append("/>\n");
 	}
 
-	public static void dimension(StringBuilder sb, int width, int height) {
+	public static void dimension(StringBuilder sb, double width, double height) {
 		// must be int otherwise parseInt("100.0") doesn't work for buttons
 		sb.append("\t<dimensions width=\"" + width + "\" height=\"" + height + "\" />\n");
 	}
